Fixes: add curlies and added type of file www.js. Refactor package for that run bin/www
