+++ package.json
@@ -9,13 +9,13 @@
         "npm": "8.3.1"
     },
     "scripts": {
-        "start": "node server.js",
-        "dev": "nodemon server.js",
+        "start": "node ./bin/www.js",
+        "dev": "nodemon ./bin/www.js",
         "build:ui": "rm -rf build && cd ../frontend && npm run build --prod && cp -r build ../backend",
         "deploy": "git push heroku master",
         "deploy:full": "npm run build:ui && git add . && git commit -m deployToHeroku && npm run deploy"
     },
-    "author": "Omiñuka Mauro",
+    "author": "Omiñuka Manu",
     "license": "ISC",
     "dependencies": {
         "@abslibs/mongoose-plugin": "^1.0.7",
@@ -37,6 +37,7 @@
         "react-paginate": "^8.0.0"
     },
     "devDependencies": {
+        "debug": "^4.3.3",
         "nodemon": "^2.0.12",
         "standard": "^16.0.4"
     },
